requirements.txt
# Core ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Experiment tracking
mlflow>=2.9.0

# API serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Monitoring & logging
evidently>=0.4.0
structlog>=23.2.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Data validation
great-expectations>=0.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# Visualization (notebooks)
matplotlib>=3.8.0
seaborn>=0.13.0

# Serialization
joblib>=1.3.0
